Move hooks to details page
